requirements.txt
altair==5.4.0
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==7.0.0
diskcache==5.6.3
distro==1.9.0
etelemetry==0.3.1
filelock==3.15.4
fitz==0.0.1.dev2
Flask==3.0.3
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.24.2
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Levenshtein==0.25.1
llama_cpp_python==0.2.83
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.4.2
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
nltk==3.8.1
numpy==2.0.1
openai==1.37.0
packaging==24.1
pandas==2.2.2
pathlib==1.0.1
pillow==10.4.0
protobuf==5.27.3
prov==2.0.1
pyarrow==17.0.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydot==3.0.1
Pygments==2.18.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.25.1
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
rapidfuzz==3.9.6
rdflib==6.3.2
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
setuptools==71.1.0
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
streamlit==1.37.1
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
torch==2.4.0
tornado==6.4.1
tqdm==4.66.4
traits==6.3.2
transformers==4.43.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
